making Minecraft actually open and partially stop duplicating recent games on main page
Games won't be added if the game is already on the recent games list. this is not a great solution as it doesn't show "recent" games like it's meant to. The best solution would be to add a game to the start of the list everytime you launch a game.